requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
channels==4.2.0
channels-redis==4.2.1
python-decouple==3.8
python-dotenv==1.1.0
colorlog==6.9.0
django-cors-headers==4.7.0
drf-yasg==1.21.10
sentry-sdk==2.24.1
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-countries==7.6.1
psycopg2-binary==2.9.10
Pillow==10.2.0 